requirements.txt
asgiref==3.8.1
certifi==2024.12.14
channels==4.2.0
charset-normalizer==3.4.0
Django==5.1.4
django-admin-inline-paginator==0.4.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
docopt==0.6.2
idna==3.10
pipreqs==0.4.13
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
yarg==0.1.10
